DRY out repeated focus / attribute logic between header cells
- EuiDataGridHeaderCell and EuiDataGridControlHeaderCell contain a bunch of repeated focus logic (that was also tweaked/fixed slightly in one place and not another a yr later - a good case study in why to DRY things out, so your bug fixes propagate)
