+++ src/components/datagrid/body/header/data_grid_control_header_cell.tsx
@@ -6,172 +6,28 @@
  * Side Public License, v 1.
  */
 
-import classnames from 'classnames';
-import React, {
-  FunctionComponent,
-  useContext,
-  useEffect,
-  useRef,
-  useState,
-} from 'react';
-import tabbable from 'tabbable';
-import { keys } from '../../../../services';
-import { DataGridFocusContext } from '../../data_grid_context';
-import { EuiDataGridControlHeaderRowProps } from '../../data_grid_types';
+import React, { FunctionComponent } from 'react';
+import { EuiDataGridControlHeaderCellProps } from '../../data_grid_types';
+import { EuiDataGridHeaderCellWrapper } from './data_grid_header_cell_wrapper';
 
-export const EuiDataGridControlHeaderCell: FunctionComponent<EuiDataGridControlHeaderRowProps> = (
-  props
-) => {
-  const { controlColumn, index, headerIsInteractive, className } = props;
-
-  const { setFocusedCell, onFocusUpdate } = useContext(DataGridFocusContext);
-
+export const EuiDataGridControlHeaderCell: FunctionComponent<EuiDataGridControlHeaderCellProps> = ({
+  controlColumn,
+  index,
+  headerIsInteractive,
+}) => {
   const { headerCellRender: HeaderCellRender, width, id } = controlColumn;
 
-  const classes = classnames('euiDataGridHeaderCell', className);
-
-  const [isFocused, setIsFocused] = useState(false);
-  useEffect(() => {
-    onFocusUpdate([index, -1], (isFocused: boolean) => {
-      setIsFocused(isFocused);
-    });
-  }, [index, onFocusUpdate]);
-
-  const headerRef = useRef<HTMLDivElement>(null);
-  const [isCellEntered, setIsCellEntered] = useState(false);
-
-  useEffect(() => {
-    if (headerRef.current) {
-      function enableInteractives() {
-        const interactiveElements = headerRef.current!.querySelectorAll(
-          '[data-euigrid-tab-managed]'
-        );
-        for (let i = 0; i < interactiveElements.length; i++) {
-          interactiveElements[i].setAttribute('tabIndex', '0');
-        }
-      }
-
-      function disableInteractives() {
-        const tababbles = tabbable(headerRef.current!);
-        if (tababbles.length > 1) {
-          console.warn(
-            `EuiDataGridHeaderCell expects at most 1 tabbable element, ${tababbles.length} found instead`
-          );
-        }
-        for (let i = 0; i < tababbles.length; i++) {
-          const element = tababbles[i];
-          element.setAttribute('data-euigrid-tab-managed', 'true');
-          element.setAttribute('tabIndex', '-1');
-        }
-      }
-
-      if (isCellEntered) {
-        enableInteractives();
-        const tabbables = tabbable(headerRef.current!);
-        if (tabbables.length > 0) {
-          tabbables[0].focus();
-        }
-      } else {
-        disableInteractives();
-      }
-    }
-  }, [isCellEntered]);
-
-  useEffect(() => {
-    if (headerRef.current) {
-      if (isFocused) {
-        const interactives = headerRef.current.querySelectorAll(
-          '[data-euigrid-tab-managed]'
-        );
-        if (interactives.length === 1) {
-          setIsCellEntered(true);
-        } else {
-          headerRef.current.focus();
-        }
-      } else {
-        setIsCellEntered(false);
-      }
-
-      // focusin bubbles while focus does not, and this needs to react to children gaining focus
-      function onFocusIn(e: FocusEvent) {
-        if (headerIsInteractive === false) {
-          // header is not interactive, avoid focusing
-          requestAnimationFrame(() => headerRef.current!.blur());
-          e.preventDefault();
-          return false;
-        } else {
-          // take the focus
-          setFocusedCell([index, -1]);
-        }
-      }
-
-      // focusout bubbles while blur does not, and this needs to react to the children losing focus
-      function onFocusOut() {
-        // wait for the next element to receive focus, then update interactives' state
-        requestAnimationFrame(() => {
-          if (headerRef.current) {
-            if (headerRef.current.contains(document.activeElement) === false) {
-              setIsCellEntered(false);
-            }
-          }
-        });
-      }
-
-      function onKeyUp(event: KeyboardEvent) {
-        switch (event.key) {
-          case keys.ENTER: {
-            event.preventDefault();
-            setIsCellEntered(true);
-            break;
-          }
-          case keys.ESCAPE: {
-            event.preventDefault();
-            // move focus to cell
-            setIsCellEntered(false);
-            headerRef.current!.focus();
-            break;
-          }
-          case keys.F2: {
-            event.preventDefault();
-            if (document.activeElement === headerRef.current) {
-              // move focus into cell's interactives
-              setIsCellEntered(true);
-            } else {
-              // move focus to cell
-              setIsCellEntered(false);
-              headerRef.current!.focus();
-            }
-            break;
-          }
-        }
-      }
-
-      const headerNode = headerRef.current;
-      // @ts-ignore-next line TS doesn't have focusin
-      headerNode.addEventListener('focusin', onFocusIn);
-      headerNode.addEventListener('focusout', onFocusOut);
-      headerNode.addEventListener('keyup', onKeyUp);
-      return () => {
-        // @ts-ignore-next line TS doesn't have focusin
-        headerNode.removeEventListener('focusin', onFocusIn);
-        headerNode.removeEventListener('focusout', onFocusOut);
-        headerNode.removeEventListener('keyup', onKeyUp);
-      };
-    }
-  }, [setFocusedCell, headerIsInteractive, isFocused, setIsCellEntered, index]);
-
   return (
-    <div
-      role="columnheader"
-      ref={headerRef}
-      tabIndex={isFocused ? 0 : -1}
-      className={classes}
-      data-test-subj={`dataGridHeaderCell-${id}`}
-      style={width != null ? { width: `${width}px` } : {}}
+    <EuiDataGridHeaderCellWrapper
+      className="euiDataGridHeaderCell--controlColumn"
+      id={id}
+      index={index}
+      width={width}
+      headerIsInteractive={headerIsInteractive}
     >
       <div className="euiDataGridHeaderCell__content">
         <HeaderCellRender />
       </div>
-    </div>
+    </EuiDataGridHeaderCellWrapper>
   );
 };
