+++ src/components/datagrid/body/header/data_grid_header_cell.tsx
@@ -11,64 +11,55 @@ import React, {
   AriaAttributes,
   FunctionComponent,
   HTMLAttributes,
-  useCallback,
   useContext,
-  useEffect,
-  useRef,
   useState,
 } from 'react';
-import tabbable from 'tabbable';
-import { keys } from '../../../../services';
 import { useGeneratedHtmlId } from '../../../../services/accessibility';
 import { EuiScreenReaderOnly } from '../../../accessibility';
 import { useEuiI18n } from '../../../i18n';
 import { EuiIcon } from '../../../icon';
 import { EuiListGroup } from '../../../list_group';
 import { EuiPopover } from '../../../popover';
-import { getColumnActions } from './column_actions';
-import { EuiDataGridColumnResizer } from './data_grid_column_resizer';
-import {
-  DataGridFocusContext,
-  DataGridSortingContext,
-} from '../../data_grid_context';
+import { DataGridSortingContext } from '../../data_grid_context';
 import { EuiDataGridHeaderCellProps } from '../../data_grid_types';
 
-export const EuiDataGridHeaderCell: FunctionComponent<EuiDataGridHeaderCellProps> = (
-  props
-) => {
-  const {
-    column,
-    index,
-    columns,
-    columnWidths,
-    schema,
-    schemaDetectors,
-    defaultColumnWidth,
-    setColumnWidth,
-    setVisibleColumns,
-    switchColumnPos,
-    headerIsInteractive,
-    className,
-  } = props;
-  const { id, display, displayAsText } = column;
+import { getColumnActions } from './column_actions';
+import { EuiDataGridColumnResizer } from './data_grid_column_resizer';
+import { EuiDataGridHeaderCellWrapper } from './data_grid_header_cell_wrapper';
 
+export const EuiDataGridHeaderCell: FunctionComponent<EuiDataGridHeaderCellProps> = ({
+  column,
+  index,
+  columns,
+  columnWidths,
+  schema,
+  schemaDetectors,
+  defaultColumnWidth,
+  setColumnWidth,
+  setVisibleColumns,
+  switchColumnPos,
+  headerIsInteractive,
+}) => {
+  const { id, display, displayAsText } = column;
   const width = columnWidths[id] || defaultColumnWidth;
 
-  const ariaProps: {
-    'aria-sort'?: AriaAttributes['aria-sort'];
-    'aria-describedby'?: AriaAttributes['aria-describedby'];
-  } = {};
+  const columnType = schema[id] ? schema[id].columnType : null;
+  const classes = classnames({
+    [`euiDataGridHeaderCell--${columnType}`]: columnType,
+  });
 
-  const screenReaderId = useGeneratedHtmlId();
-  let sortString;
   const actionButtonAriaLabel = useEuiI18n(
     'euiDataGridHeaderCell.headerActions',
     'Header actions'
   );
+  const ariaProps: {
+    'aria-sort'?: AriaAttributes['aria-sort'];
+    'aria-describedby'?: AriaAttributes['aria-describedby'];
+  } = {};
+  const screenReaderId = useGeneratedHtmlId();
 
   const sorting = useContext(DataGridSortingContext);
-  const { setFocusedCell, onFocusUpdate } = useContext(DataGridFocusContext);
-
+  let sortString;
   if (sorting) {
     const sortedColumnIds = new Set(sorting.columns.map(({ id }) => id));
 
@@ -91,169 +82,7 @@ export const EuiDataGridHeaderCell: FunctionComponent<EuiDataGridHeaderCellProps
     }
   }
 
-  const columnType = schema[id] ? schema[id].columnType : null;
-
-  const classes = classnames(
-    'euiDataGridHeaderCell',
-    {
-      [`euiDataGridHeaderCell--${columnType}`]: columnType,
-    },
-    className
-  );
-
-  const [isFocused, setIsFocused] = useState(false);
-  useEffect(() => {
-    return onFocusUpdate([index, -1], (isFocused: boolean) => {
-      setIsFocused(isFocused);
-    });
-  }, [index, onFocusUpdate]);
-
-  const headerRef = useRef<HTMLDivElement>(null);
-  const [isCellEntered, setIsCellEntered] = useState(false);
   const [isPopoverOpen, setIsPopoverOpen] = useState(false);
-
-  const enableInteractives = useCallback(() => {
-    if (headerRef.current) {
-      const interactiveElements = headerRef.current.querySelectorAll(
-        '[data-euigrid-tab-managed]'
-      );
-      for (let i = 0; i < interactiveElements.length; i++) {
-        interactiveElements[i].setAttribute('tabIndex', '0');
-      }
-    }
-  }, []);
-
-  const disableInteractives = useCallback(() => {
-    if (headerRef.current) {
-      const tababbles = tabbable(headerRef.current);
-      if (tababbles.length > 1) {
-        console.warn(
-          `EuiDataGridHeaderCell expects at most 1 tabbable element, ${tababbles.length} found instead`
-        );
-      }
-      for (let i = 0; i < tababbles.length; i++) {
-        const element = tababbles[i];
-        element.setAttribute('data-euigrid-tab-managed', 'true');
-        element.setAttribute('tabIndex', '-1');
-      }
-    }
-  }, []);
-
-  useEffect(() => {
-    if (headerRef.current) {
-      if (isCellEntered) {
-        enableInteractives();
-        const tabbables = tabbable(headerRef.current!);
-        if (tabbables.length > 0) {
-          tabbables[0].focus();
-        }
-      } else {
-        disableInteractives();
-      }
-    }
-  }, [disableInteractives, enableInteractives, isCellEntered]);
-
-  useEffect(() => {
-    if (headerRef.current) {
-      if (isFocused) {
-        const interactives = headerRef.current.querySelectorAll(
-          '[data-euigrid-tab-managed]'
-        );
-        if (interactives.length === 1) {
-          setIsCellEntered(true);
-        } else {
-          headerRef.current.focus();
-        }
-      } else {
-        setIsCellEntered(false);
-      }
-
-      // focusin bubbles while focus does not, and this needs to react to children gaining focus
-      function onFocusIn(e: FocusEvent) {
-        if (!headerIsInteractive) {
-          // header is not interactive, avoid focusing
-          requestAnimationFrame(() => headerRef.current!.blur());
-          e.preventDefault();
-          return false;
-        } else {
-          // take the focus
-          if (isFocused === false) {
-            setFocusedCell([index, -1]);
-          } else if (headerRef.current) {
-            // this cell already had the grid's focus, so re-enable interactives
-            enableInteractives();
-            setIsCellEntered(true);
-
-            // if there is only one interactive element shift focus to the interactive element
-            const tabbables = tabbable(headerRef.current);
-            if (tabbables.length === 1) {
-              tabbables[0].focus();
-            }
-          }
-        }
-      }
-
-      // focusout bubbles while blur does not, and this needs to react to the children losing focus
-      function onFocusOut() {
-        // wait for the next element to receive focus, then update interactives' state
-        requestAnimationFrame(() => {
-          if (headerRef.current) {
-            if (!headerRef.current.contains(document.activeElement)) {
-              setIsCellEntered(false);
-            }
-          }
-        });
-      }
-
-      function onKeyUp(event: KeyboardEvent) {
-        switch (event.key) {
-          case keys.ENTER: {
-            event.preventDefault();
-            setIsCellEntered(true);
-            break;
-          }
-          case keys.ESCAPE: {
-            event.preventDefault();
-            // move focus to cell
-            setIsCellEntered(false);
-            headerRef.current!.focus();
-            break;
-          }
-          case keys.F2: {
-            event.preventDefault();
-            if (document.activeElement === headerRef.current) {
-              // move focus into cell's interactives
-              setIsCellEntered(true);
-            } else {
-              // move focus to cell
-              setIsCellEntered(false);
-              headerRef.current!.focus();
-            }
-            break;
-          }
-        }
-      }
-
-      const headerNode = headerRef.current;
-      // @ts-ignore-next line TS doesn't have focusin
-      headerNode.addEventListener('focusin', onFocusIn);
-      headerNode.addEventListener('focusout', onFocusOut);
-      headerNode.addEventListener('keyup', onKeyUp);
-      return () => {
-        // @ts-ignore-next line TS doesn't have focusin
-        headerNode.removeEventListener('focusin', onFocusIn);
-        headerNode.removeEventListener('focusout', onFocusOut);
-        headerNode.removeEventListener('keyup', onKeyUp);
-      };
-    }
-  }, [
-    enableInteractives,
-    headerIsInteractive,
-    isFocused,
-    setIsCellEntered,
-    index,
-    setFocusedCell,
-  ]);
 
   const columnActions = getColumnActions(
     column,
@@ -278,14 +107,13 @@ export const EuiDataGridHeaderCell: FunctionComponent<EuiDataGridHeaderCellProps
   ) : null;
 
   return (
-    <div
-      role="columnheader"
-      {...ariaProps}
-      ref={headerRef}
-      tabIndex={isFocused && !isCellEntered ? 0 : -1}
+    <EuiDataGridHeaderCellWrapper
+      id={id}
+      index={index}
+      width={width}
+      headerIsInteractive={headerIsInteractive}
       className={classes}
-      data-test-subj={`dataGridHeaderCell-${id}`}
-      style={width != null ? { width: `${width}px` } : {}}
+      {...ariaProps}
     >
       {column.isResizable !== false && width != null ? (
         <EuiDataGridColumnResizer
@@ -345,6 +173,6 @@ export const EuiDataGridHeaderCell: FunctionComponent<EuiDataGridHeaderCellProps
           </div>
         </EuiPopover>
       )}
-    </div>
+    </EuiDataGridHeaderCellWrapper>
   );
 };
