+++ src/components/datagrid/body/header/data_grid_header_row.tsx
@@ -50,7 +50,6 @@ const EuiDataGridHeaderRow = forwardRef<
           index={index}
           controlColumn={controlColumn}
           headerIsInteractive={headerIsInteractive}
-          className="euiDataGridHeaderCell--controlColumn"
         />
       ))}
       {columns.map((column, index) => (
@@ -75,7 +74,6 @@ const EuiDataGridHeaderRow = forwardRef<
           index={index + leadingControlColumns.length + columns.length}
           controlColumn={controlColumn}
           headerIsInteractive={headerIsInteractive}
-          className="euiDataGridHeaderCell--controlColumn"
         />
       ))}
     </div>
